Add ownership verification status fields
Part of requesting a custom hostname from Cloudflare involves proving
that you are the owner of the domain. This is in addition to the
verification that occurs around the SSL certificate. We were exposing
the fields that contained SSL verification information but not the ones
exposing domain verification. This adds those fields under a new
`ownershipVerification` entry in the status of the resource.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>