Make sure Zone VanityNS is compared in any order
Previously, we were relying on the Cloudflare API to return our VanityNS
field in the same order as we submitted it, otherwise we would
constantly return `UpToDate`: false. It looks like this _was_ the case,
either through luck or design, but by adding `cmpopts.EquateEmpty()` and
our own `cmpopts.SortSlices` function, we can avoid changes in the API
causing us issues down the line.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>